Split ffi_types.rs into focused sub-modules (#596)
Replace the monolithic ffi_types.rs (~1,187 lines) with 7 focused
sub-modules under src/ffi_types/:

- creature.rs: Creature, neuron, synapse JSON representations
- requests.rs: FFI request structs (input from NEAT-AI)
- responses.rs: FFI response structs (output to NEAT-AI)
- candidates.rs: Candidate-related types
- session.rs: Streaming session types
- diagnostics.rs: Diagnostic, metadata, GPU info, timing types
- conversions.rs: Conversion helpers (internal → JSON)
- mod.rs: Public API, re-exports for backward compatibility

All public types are re-exported at the ffi_types:: level,
preserving full backward compatibility. No consumer code changes
required.
