+++ src/ffi_types/creature.rs
@@ -1,0 +1,66 @@
+//! Creature, neuron, and synapse JSON representations for the FFI boundary.
+
+use serde::{Deserialize, Serialize};
+
+// ============================================================================
+// Creature / Neuron / Synapse representations
+// ============================================================================
+
+/// JSON representation of Creature
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct CreatureJson {
+    pub neurons: Vec<NeuronJson>,
+    pub synapses: Vec<SynapseJson>,
+    pub input: usize,
+    pub output: usize,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct NeuronJson {
+    pub uuid: String,
+    #[serde(rename = "type")]
+    pub neuron_type: String,
+    /// Activation function. Defaults to "IDENTITY" for constant neurons.
+    #[serde(default = "default_squash")]
+    pub squash: String,
+    #[serde(default)]
+    pub bias: f32,
+}
+
+fn default_squash() -> String {
+    "IDENTITY".to_string()
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone, Default)]
+pub struct SynapseJson {
+    #[serde(default, alias = "fromUUID")]
+    pub from_uuid: String,
+    #[serde(default, alias = "toUUID")]
+    pub to_uuid: String,
+    #[serde(default)]
+    pub weight: f32,
+    /// Synapse type for IF neurons: "condition", "positive", or "negative".
+    /// Used to determine which synapses contribute to condition evaluation
+    /// versus positive/negative branches. None for non-IF neurons.
+    #[serde(default, rename = "type")]
+    pub synapse_type: Option<String>,
+}
+
+/// Pre-computed neuron data for a single neuron
+#[derive(Debug, Deserialize, Clone)]
+pub struct NeuronData {
+    pub neuron_uuid: String,
+    pub activation: f32,
+    #[serde(default)]
+    pub value: Option<f32>,
+    pub errors: Vec<f32>,
+}
+
+/// Training data record
+#[derive(Debug, Deserialize, Clone)]
+pub struct TrainingRecord {
+    pub input: Vec<f32>,
+    pub output: Vec<f32>,
+    #[serde(default)]
+    pub neuron_data: Option<Vec<NeuronData>>,
+}
