+++ src/ffi_types/diagnostics.rs
@@ -1,0 +1,266 @@
+//! Diagnostic, metadata, GPU info, and timing types for the FFI boundary.
+
+use serde::Serialize;
+
+use crate::analysis;
+
+// =============================================================================
+// Neuron stats
+// =============================================================================
+
+#[derive(Debug, Serialize, Clone, Copy)]
+#[serde(rename_all = "camelCase")]
+pub struct NeuronStatsJson {
+    pub mean_error: f32,
+    pub error_variance: f32,
+    pub mean_activation: f32,
+    pub activation_variance: f32,
+    pub error_spike_count: u32,
+    pub activation_spike_count: u32,
+    pub activation_min: f32,
+    pub activation_max: f32,
+}
+
+// =============================================================================
+// Analysis metadata
+// =============================================================================
+
+/// JSON representation of synapse analysis metadata.
+///
+/// Surfaces diagnostic information to help callers understand:
+/// - Whether prediction accuracy is degraded (missing `value` data)
+/// - Whether candidates were truncated by `maxCandidates`
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SynapseAnalysisMetadataJson {
+    /// Whether `targetValue` (pre-activation) was available in the recorded data.
+    pub target_value_available: bool,
+    /// Whether saturation-aware simulation was used for at least one candidate.
+    pub saturation_aware_simulation_used: bool,
+    /// Total number of synapse candidates found during analysis (before truncation).
+    pub candidates_found: usize,
+    /// Number of synapse candidates returned to caller (after `maxCandidates` truncation).
+    pub candidates_returned: usize,
+    /// True when analysis hit its deadline and returned partial results.
+    pub timed_out: bool,
+    /// Number of focus neurons completed before returning.
+    pub completed_focus_neurons: usize,
+    /// Total focus neurons requested for this analysis invocation.
+    pub total_focus_neurons: usize,
+    /// Minimum input index observed with non-empty records (eg 0).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub input_index_min_seen_with_records: Option<usize>,
+    /// Maximum input index observed with non-empty records (eg 1555).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub input_index_max_seen_with_records: Option<usize>,
+    /// GPU timing data for performance diagnostics (Issue #195).
+    /// Only present when `NEAT_AI_DISCOVERY_GPU_TIMING=1` is set.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub timing: Option<AnalysisTimingJson>,
+    /// Information about the GPU adapter used (Issue #228).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub gpu_info: Option<GpuAdapterInfoJson>,
+    /// Per-discovery-module statistics for the current run (Issue #485).
+    /// Reports how many candidates each module produced and historical success rates.
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    pub discovery_module_stats: Vec<analysis::module_weights::DiscoveryModuleStatsJson>,
+}
+
+/// JSON representation of neuron analysis metadata.
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct NeuronAnalysisMetadataJson {
+    /// Total number of neuron candidates found during analysis (before truncation).
+    pub candidates_found: usize,
+    /// Number of neuron candidates returned to caller (after `maxCandidates` truncation).
+    pub candidates_returned: usize,
+    /// True when analysis hit its deadline and returned partial results.
+    pub timed_out: bool,
+    /// Number of focus neurons completed before returning.
+    pub completed_focus_neurons: usize,
+    /// Total focus neurons requested for this analysis invocation.
+    pub total_focus_neurons: usize,
+    /// GPU timing data for performance diagnostics (Issue #195).
+    /// Only present when `NEAT_AI_DISCOVERY_GPU_TIMING=1` is set.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub timing: Option<AnalysisTimingJson>,
+    /// Information about the GPU adapter used (Issue #228).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub gpu_info: Option<GpuAdapterInfoJson>,
+}
+
+// =============================================================================
+// GPU Info JSON Types (Issue #228)
+// =============================================================================
+
+/// JSON representation of GPU adapter information.
+#[derive(Debug, Serialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct GpuAdapterInfoJson {
+    /// Human-readable name of the GPU (e.g., "Apple M4 Pro").
+    pub name: String,
+    /// Whether the GPU has unified memory architecture.
+    pub unified_memory: bool,
+    /// Whether zero-copy buffer sharing is currently enabled.
+    pub zero_copy_enabled: bool,
+}
+
+// =============================================================================
+// GPU Timing JSON Types (Issue #195)
+// =============================================================================
+
+/// JSON representation of per-shader timing statistics.
+#[derive(Debug, Serialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct ShaderTimingJson {
+    /// Number of times this shader was executed.
+    pub calls: u32,
+    /// Total execution time in milliseconds.
+    pub total_ms: f64,
+    /// Average execution time per call in milliseconds.
+    pub avg_ms: f64,
+}
+
+/// JSON representation of GPU-side timing breakdown.
+#[derive(Debug, Serialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct GpuTimingBreakdownJson {
+    /// Total time spent in shader execution (all shaders combined) in milliseconds.
+    pub shader_execution_ms: f64,
+    /// Total time spent in buffer mapping/transfers in milliseconds.
+    pub buffer_transfer_ms: f64,
+    /// Per-shader timing statistics.
+    /// Keys are shader names: "helpful", "harmful", "relu", "activation", "bias"
+    #[serde(skip_serializing_if = "std::collections::HashMap::is_empty")]
+    pub shader_timings: std::collections::HashMap<String, ShaderTimingJson>,
+}
+
+/// JSON representation of CPU-side timing breakdown.
+#[derive(Debug, Serialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct CpuTimingBreakdownJson {
+    /// Time spent building samples for GPU evaluation in milliseconds.
+    pub sample_building_ms: f64,
+    /// Time spent processing results from GPU in milliseconds.
+    pub result_processing_ms: f64,
+}
+
+/// JSON representation of complete timing data for an analysis run.
+///
+/// Only populated when `NEAT_AI_DISCOVERY_GPU_TIMING=1` is set.
+#[derive(Debug, Serialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct AnalysisTimingJson {
+    /// Total wall-clock time for the analysis in milliseconds.
+    pub total_analysis_ms: f64,
+    /// GPU-side timing breakdown.
+    pub gpu: GpuTimingBreakdownJson,
+    /// CPU-side timing breakdown.
+    pub cpu: CpuTimingBreakdownJson,
+}
+
+// =============================================================================
+// Synapse diagnostics
+// =============================================================================
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SynapseDiagnosticJson {
+    pub target_neuron_uuid: String,
+    pub reason: SynapseDiagnosticReasonJson,
+    pub evaluated_candidates: u32,
+    pub candidates_with_samples: u32,
+    pub target_record_count: usize,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub detail: Option<SynapseDiagnosticDetailJson>,
+}
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "snake_case")]
+pub enum SynapseDiagnosticReasonJson {
+    NoEligibleSources,
+    NoDiagnostics,
+    NoSamples,
+    ZeroImprovement,
+    BelowThreshold,
+}
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SynapseDiagnosticDetailJson {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub source_neuron_uuid: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub sample_count: Option<usize>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub source_record_count: Option<usize>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub improved_count: Option<u32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub worsened_count: Option<u32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub expected_creature_error_reduction: Option<f32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub threshold: Option<f32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub suggested_weight: Option<f32>,
+}
+
+// =============================================================================
+// Neuron diagnostics
+// =============================================================================
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct NeuronDiagnosticJson {
+    pub target_neuron_uuid: String,
+    pub reason: NeuronDiagnosticReasonJson,
+    pub evaluated_sources: u32,
+    pub sources_with_samples: u32,
+    pub target_record_count: usize,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub detail: Option<NeuronDiagnosticDetailJson>,
+}
+
+#[derive(Debug, Clone, Serialize)]
+#[serde(rename_all = "snake_case")]
+pub enum NeuronDiagnosticReasonJson {
+    NoEligibleSources,
+    NoDiagnostics,
+    NoSamples,
+    NotEnoughActivations,
+    WeightDegenerate,
+    BelowThreshold,
+    /// Hidden neurons are filtered out from add-neuron analysis because their
+    /// backpropagated errors don't reliably translate to output error reduction.
+    HiddenNeuronFiltered,
+    /// Input neurons are filtered out from add-neuron analysis because they're
+    /// observation sources, not computation nodes - they have no activation function
+    /// or error to reduce.
+    InputNeuronFiltered,
+    /// Constant neurons are filtered out from add-neuron analysis because they
+    /// don't receive inputs - they always output a fixed value regardless of
+    /// network state, so adding a connection to them has no effect.
+    ConstantNeuronFiltered,
+}
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct NeuronDiagnosticDetailJson {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub source_neuron_uuid: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub orientation: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub sample_count: Option<usize>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub improved_count: Option<u32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub worsened_count: Option<u32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub expected_creature_error_reduction: Option<f32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub threshold: Option<f32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub outgoing_weight: Option<f32>,
+}
