+++ src/ffi_types/responses.rs
@@ -1,0 +1,181 @@
+//! FFI response structs (output to NEAT-AI).
+
+use serde::Serialize;
+
+use crate::analysis;
+
+use super::candidates::{
+    CandidateNeuronJson, CandidateSynapseJson, CoordinatedStructuralCandidateJson,
+    RankedNeuronJson, RemovalCandidateJson, SynapseWeightUpdateCandidateJson,
+};
+use super::diagnostics::{
+    NeuronAnalysisMetadataJson, NeuronDiagnosticJson, SynapseAnalysisMetadataJson,
+    SynapseDiagnosticJson,
+};
+
+/// JSON output from record_discovery function
+#[derive(Debug, Serialize)]
+pub struct RecordDiscoveryOutput {
+    pub success: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub temp_dir: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub file: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<String>,
+}
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct AnalyzeParallelOutput {
+    pub success: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub helpful_synapses: Option<Vec<CandidateSynapseJson>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub harmful_synapses: Option<Vec<CandidateSynapseJson>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub synapse_diagnostics: Option<Vec<SynapseDiagnosticJson>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub synapse_gpu_used: Option<bool>,
+    /// Synapse analysis metadata for observability (v0.2.17+).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub synapse_metadata: Option<SynapseAnalysisMetadataJson>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub helpful_neurons: Option<Vec<CandidateNeuronJson>>,
+    /// Candidates to update weights on existing synapses (v0.2.18+).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub synapse_weight_updates: Option<Vec<SynapseWeightUpdateCandidateJson>>,
+    /// Coordinated (grouped) structural candidates (v0.2.18+).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub coordinated_structural_candidates: Option<Vec<CoordinatedStructuralCandidateJson>>,
+    /// Candidate clusters for redundancy reduction (Issue #224).
+    ///
+    /// Groups similar candidates by target neuron so the controller can test a
+    /// representative first and skip redundant ablation tests.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub candidate_clusters: Option<Vec<analysis::candidate_clustering::CandidateClusterJson>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub neuron_diagnostics: Option<Vec<NeuronDiagnosticJson>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub neuron_gpu_used: Option<bool>,
+    /// Neuron analysis metadata for observability (v0.2.17+).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub neuron_metadata: Option<NeuronAnalysisMetadataJson>,
+    /// Current neuron fingerprints for incremental analysis (Issue #490).
+    ///
+    /// Callers should store these and pass them back as `previousNeuronFingerprints`
+    /// on the next discovery run to enable incremental analysis.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub neuron_fingerprints:
+        Option<std::collections::HashMap<String, analysis::neuron_fingerprint::NeuronFingerprint>>,
+    /// Number of focus neurons skipped due to unchanged fingerprints (Issue #490).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub fingerprint_cache_hits: Option<usize>,
+    /// Number of focus neurons analysed (changed or new fingerprints) (Issue #490).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub fingerprint_cache_misses: Option<usize>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<String>,
+}
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct CheckGpuOutput {
+    pub success: bool,
+    pub gpu_available: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub reason: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<String>,
+}
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GetVersionOutput {
+    pub success: bool,
+    pub version: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<String>,
+}
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct RankFocusNeuronsOutput {
+    pub success: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub neurons: Option<Vec<RankedNeuronJson>>,
+    /// Neurons with high error but very low impact - candidates for removal
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub removal_candidates: Option<Vec<RemovalCandidateJson>>,
+    /// Issue #306: Coordinated structural candidates for removing constant-value neurons.
+    /// When a hidden neuron has near-zero activation variance (constant output), it can be
+    /// removed and its effect folded into bias adjustments for downstream neurons.
+    /// Each candidate contains:
+    /// - A RemoveNeuron operation for the constant neuron
+    /// - SetBias operations for all downstream neurons with adjusted biases
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub constant_neuron_removals: Option<Vec<CoordinatedStructuralCandidateJson>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub max_output_error: Option<f32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub processed_neurons: Option<usize>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub total_neurons: Option<usize>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub duration_ms: Option<u64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<String>,
+}
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct MergeParquetOutput {
+    pub success: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub output_file: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<String>,
+}
+
+/// JSON output from export_visualisation_snapshot function
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ExportVisualisationSnapshotOutput {
+    pub success: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub out_file: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub stats: Option<ExportVisualisationStats>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<String>,
+}
+
+/// Statistics from the export operation
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ExportVisualisationStats {
+    pub obs_count: usize,
+    pub neuron_count: usize,
+    pub synapse_count: usize,
+    pub output_count: usize,
+}
+
+/// JSON output from read_discovery_records function
+#[derive(Debug, Serialize)]
+pub struct ReadDiscoveryOutput {
+    pub success: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub records: Option<Vec<DiscoverRecordJson>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<String>,
+}
+
+/// JSON representation of DiscoverRecord for serialization
+#[derive(Debug, Serialize)]
+pub struct DiscoverRecordJson {
+    pub obs_index: u32,
+    pub neuron_uuid: String,
+    pub value: Option<f32>,
+    pub activation: f32,
+    pub errors: Vec<f32>,
+}
